+++ src/useConnectCall.test.ts
@@ -24,7 +24,6 @@ const onPeerConnected = jest.fn();
 const onPeerDisconnected = jest.fn();
 const onNewMessage = jest.fn();
 const onTimer = jest.fn();
-const onConnectionState = jest.fn();
 
 const call = {
   id: "2",
@@ -48,7 +47,6 @@ describe("useConnectCall", () => {
     onPeerConnected.mockClear();
     onPeerDisconnected.mockClear();
     onNewMessage.mockClear();
-    onConnectionState.mockClear();
   });
 
   it("completes the connection handshake", async () => {
@@ -59,7 +57,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
 
@@ -76,7 +73,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
 
@@ -105,7 +101,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
 
@@ -135,7 +130,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
 
@@ -156,7 +150,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
 
@@ -171,6 +164,10 @@ describe("useConnectCall", () => {
     expect(result.current.peers).toMatchInlineSnapshot(`
       Array [
         Object {
+          "connectionState": Object {
+            "ping": NaN,
+            "quality": "unknown",
+          },
           "stream": MediaStream {
             "tracks": Array [
               Object {
@@ -196,6 +193,10 @@ describe("useConnectCall", () => {
     expect(result.current.peers).toMatchInlineSnapshot(`
       Array [
         Object {
+          "connectionState": Object {
+            "ping": NaN,
+            "quality": "unknown",
+          },
           "stream": MediaStream {
             "tracks": Array [
               Object {
@@ -228,6 +229,10 @@ describe("useConnectCall", () => {
     expect(result.current.peers).toMatchInlineSnapshot(`
       Array [
         Object {
+          "connectionState": Object {
+            "ping": NaN,
+            "quality": "unknown",
+          },
           "stream": MediaStream {
             "tracks": Array [
               Object {
@@ -259,7 +264,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
     await waitFor(() => expect(result.current.status).toBe("connected"));
@@ -290,7 +294,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
     await waitFor(() => expect(result.current.status).toBe("connected"));
@@ -308,7 +311,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
     await waitFor(() => expect(result.current.status).toBe("connected"));
@@ -344,7 +346,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
     await waitFor(() => expect(result.current.status).toBe("connected"));
@@ -373,7 +374,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );
     await waitFor(() => expect(result.current.status).toBe("connected"));
@@ -402,7 +402,6 @@ describe("useConnectCall", () => {
         onPeerConnected,
         onPeerDisconnected,
         onNewMessage,
-        onConnectionState,
       })
     );